move: 상태 비교 함수를 states 폴더로 이동
